+++ src/javascript/utils/directory.beautifier.js
@@ -1,0 +1,8 @@
+import requirePath from './require'
+const path = requirePath("path")
+
+function beautifyDir(dir){
+	return path.normalize(dir).replace(/\\\\/g,'\\').replace(/"/g,'')
+}
+
+export default beautifyDir